requirements.txt
transformers
sentence-transformers
numpy
pyyaml
fastapi
uvicorn
beautifulsoup4
requests
